# apps/api/cmd/server/openapi.bundled.yaml
swagger: '2.0'
info:
  title: Terp API
  description: |
    RESTful API for the Terp application - Time & Resource Planning.

    ## Authentication
    All endpoints except `/health` and `/auth/*` require a valid JWT token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```

    ## Pagination
    List endpoints support cursor-based pagination:
    - `limit`: Number of items to return (default: 20, max: 100)
    - `cursor`: Cursor for next page (returned in response)

    ## Error Responses
    All errors follow RFC 7807 Problem Details format.

    ## File Structure
    This spec is organized into:
    - `paths/` - API endpoint definitions by domain
    - `schemas/` - Data models by domain
    - `responses/` - Reusable response definitions
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: Proprietary
host: 'localhost:8080'
basePath: /api/v1
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: JWT token. Format "Bearer <token>"
security:
  - BearerAuth: []
paths:
...
  '/monthly-values/{id}/close':
    post:
      tags:
        - Monthly Values
      summary: Close month
      description: 'Closes the month for the employee, preventing further modifications'
      operationId: closeMonth
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
        - name: body
          in: body
          schema:
            type: object
            properties:
              recalculate:
                type: boolean
                description: Recalculate all daily values before closing
                default: true
              notes:
                type: string
      responses:
        '200':
          description: Month closed
          schema:
            $ref: '#/definitions/MonthlyValue'
        '400':
          description: Cannot close (has errors or already closed)
          schema:
            $ref: '#/definitions/ProblemDetails'
        '401':
          $ref: '#/responses/Unauthorized'
        '404':
          $ref: '#/responses/NotFound'
  '/monthly-values/{id}/reopen':
    post:
      tags:
        - Monthly Values
      summary: Reopen month
      description: Reopens a closed month for corrections
      operationId: reopenMonth
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
        - name: body
          in: body
          required: true
          schema:
            type: object
            required:
              - reason
            properties:
              reason:
                type: string
                minLength: 10
                description: Reason for reopening
      responses:
        '200':
          description: Month reopened
          schema:
            $ref: '#/definitions/MonthlyValue'
        '400':
          description: Cannot reopen (not closed or already exported)
          schema:
            $ref: '#/definitions/ProblemDetails'
        '401':
          $ref: '#/responses/Unauthorized'
        '404':
          $ref: '#/responses/NotFound'
...
definitions:
...
  ProblemDetails:
    type: object
    required:
      - type
      - title
      - status
    properties:
      type:
        type: string
        format: uri
        description: URI identifying the problem type
        example: 'https://api.example.com/problems/validation-error'
      title:
        type: string
        description: Short human-readable summary
        example: Validation Error
      status:
        type: integer
        description: HTTP status code
        example: 400
      detail:
        type: string
        description: Human-readable explanation
        example: The 'content' field is required
      instance:
        type: string
        format: uri
        description: URI identifying the specific occurrence
      errors:
        type: array
        description: Field-level validation errors
        items:
          type: object
          properties:
            field:
              type: string
              example: content
            message:
              type: string
              example: Field is required
...
  EmployeeSummary:
    type: object
    required:
      - id
      - personnel_number
      - first_name
      - last_name
    properties:
      id:
        type: string
        format: uuid
      personnel_number:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      department_id:
        type: string
        format: uuid
        x-nullable: true
      tariff_id:
        type: string
        format: uuid
        x-nullable: true
      is_active:
        type: boolean
...
  MonthlyValue:
    type: object
    required:
      - id
      - tenant_id
      - employee_id
      - year
      - month
    properties:
      id:
        type: string
        format: uuid
      tenant_id:
        type: string
        format: uuid
      employee_id:
        type: string
        format: uuid
      year:
        type: integer
        example: 2024
      month:
        type: integer
        minimum: 1
        maximum: 12
        example: 1
      status:
        type: string
        enum:
          - open
          - calculated
          - closed
          - exported
        example: calculated
      target_minutes:
        type: integer
        description: Total target work time for month
        example: 10080
      gross_minutes:
        type: integer
        description: Total gross work time
        example: 10500
      break_minutes:
        type: integer
        description: Total break time
        example: 600
      net_minutes:
        type: integer
        description: Total net work time
        example: 9900
      overtime_minutes:
        type: integer
        description: Total overtime
        example: 420
      undertime_minutes:
        type: integer
        description: Total undertime
        example: 0
      balance_minutes:
        type: integer
        description: Balance (overtime - undertime)
        example: 420
      working_days:
        type: integer
        description: Number of working days in month
        example: 21
      worked_days:
        type: integer
        description: Number of days actually worked
        example: 20
      absence_days:
        type: number
        format: decimal
        description: Number of absence days
        example: 1
      holiday_days:
        type: integer
        description: Number of holidays
        example: 0
      account_balances:
        type: object
        description: Balances per account (account_id -> minutes)
        additionalProperties:
          type: integer
      calculated_at:
        type: string
        format: date-time
        x-nullable: true
      closed_at:
        type: string
        format: date-time
        x-nullable: true
      closed_by:
        type: string
        format: uuid
        x-nullable: true
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time
      employee:
        allOf:
          - $ref: '#/definitions/EmployeeSummary'
        x-nullable: true
...
responses:
...
  Unauthorized:
    description: Missing or invalid authentication
    schema:
      $ref: '#/definitions/ProblemDetails'
...
  NotFound:
    description: Resource not found
    schema:
      $ref: '#/definitions/ProblemDetails'
...